+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## XSTAMPP 3.0.0
+
+#### A-STPA Bugfixes:
+
+*  Fix multiple User responsibilities not working when more than one user is selected
+*  Fix in Causal Factors table the Notes / Rationale is not imported in v2.6
+*  Fix in Causal Factors table the design hints for single safety constraint rows overwrite the notes
+
 ## XSTAMPP 2.6.0
 
 #### A-STPA Features: